moves CustomerInfoResponseHandler
