+++ common/src/test/java/com/revenuecat/purchases/common/offlineentitlements/CustomerInfoResponseHandlerTest.kt
@@ -1,0 +1,4 @@
+package com.revenuecat.purchases.common.offlineentitlements
+
+class CustomerInfoResponseHandlerTest {
+}
